updated validation helper function to output 1 and 0 instead of true and false

+++ projects/_01_ingest/cqc_api/utils/validate_cqc_locations.py
@@ -74,11 +74,17 @@ def add_list_column_validation_check_flags(
 
     for col in columns:
         validation_expr = (
-            pl.col(col).is_null()
-            | ((pl.col(col).list.len() > 0) & (~pl.col(col).list.contains(None)))
-        ).alias(f"{col}_has_no_empty_or_null")
+            (
+                pl.col(col).is_null()
+                | ((pl.col(col).list.len() > 0) & (~pl.col(col).list.contains(None)))
+            )
+            .cast(pl.Int8)
+            .alias(f"{col}_has_no_empty_or_null")
+        )
 
-        completeness_expr = pl.col(col).is_not_null().alias(f"{col}_is_not_null")
+        completeness_expr = (
+            pl.col(col).is_not_null().cast(pl.Int8).alias(f"{col}_is_not_null")
+        )
 
         expressions.extend([validation_expr, completeness_expr])
